refactor: convert exit command to jsx component with app store integration
